move provider mocks to providers folder
